rename script file, add logic to update timestamp and recommendation co2esaving value
